♻️ refactor: move PublishMessage to messages.go and update queue handlers to support vhost parameter
